Sync md and docx

+++ listings/ch11-writing-automated-tests/listing-11-07/src/lib.rs
@@ -1,4 +1,4 @@
-pub fn add_two(a: usize) -> usize {
+pub fn add_two(a: u64) -> u64 {
     a + 2
 }
